feat(qfc): Improve Quickfix auto close timer management
This commit introduces several changes to the Quickfix auto close timer
management in the qfc module. The default timeout has been increased from
2000 to 3000 milliseconds to provide a longer window before the Quickfix
window is automatically closed.

A new function, `stop_qf_timer`, has been added to stop the timer when
necessary. This function is called when the 'WinEnter' autocommand is
triggered, effectively stopping the timer when the Quickfix window is
entered. This prevents the Quickfix window from closing while it is being
used.

The 'WinLeave' autocommand now only starts the timer if the buffer type is
'quickfix', ensuring that the timer is only active when it is needed.

These changes improve the user experience by making the auto close behavior
more intuitive and less disruptive.